modified rake task so it won't reset the database in production

+++ lib/tasks/populate.rake
@@ -19,10 +19,10 @@ def cyan(text);     colorize(text, 36); end
 namespace :db do
   desc "Fill database with sample data" 
   task populate: :environment do
-    # if Rails.env.development?
+    if Rails.env.development?
       puts "#{red("==>")} Clearing Current Data"
       Rake::Task['db:reset'].invoke
-    # end
+    end
     # puts "#{green("==>")} Making sample users"
     # make_users
     # puts "#{green("==>")} Making user roles"
